Q4: Letter Conversion
This code asks the user to enter a character. Then the program asks the user to enter a nonnegative
integer offset. This program adds the offset to the character to produce a new ASCII value. This
value should then be displayed. If the offset is 0 and the character is a letter, the program should change
the case of the letter. If the offset is 0 but the character is not a letter, the same character should be displayed.