Q5: Warmer or Colder?
This program helps you guess a random integer number. Initially the program creates a random
number. Then the user inputs a guess. After the first guess, the computer
says whether the new guess is warmer (closer) or colder (farther) from the last guess (absolute value of
the difference). In the event that the distance is the same it should say “No change”.